Minor code rearranging
