+++ app/src/main/java/com/javernaut/whatthecodec/home/data/model/AudioStreamFeature.kt
@@ -1,0 +1,15 @@
+package com.javernaut.whatthecodec.home.data.model
+
+import androidx.annotation.Keep
+
+@Keep
+enum class AudioStreamFeature {
+    Codec,
+    Bitrate,
+    Channels,
+    ChannelLayout,
+    SampleFormat,
+    SampleRate,
+    Language,
+    Disposition
+}
