+++ app/src/main/java/com/javernaut/whatthecodec/home/data/model/SubtitleStreamFeature.kt
@@ -1,0 +1,10 @@
+package com.javernaut.whatthecodec.home.data.model
+
+import androidx.annotation.Keep
+
+@Keep
+enum class SubtitleStreamFeature {
+    Codec,
+    Language,
+    Disposition
+}
