+++ app/src/main/java/com/javernaut/whatthecodec/home/data/model/VideoStreamFeature.kt
@@ -1,0 +1,14 @@
+package com.javernaut.whatthecodec.home.data.model
+
+import androidx.annotation.Keep
+
+@Keep
+enum class VideoStreamFeature {
+    Codec,
+    Bitrate,
+    FrameRate,
+    FrameWidth,
+    FrameHeight,
+    Language,
+    Disposition
+}
