+++ app/src/main/java/com/javernaut/whatthecodec/home/presentation/model/ScreenMessage.kt
@@ -1,0 +1,7 @@
+package com.javernaut.whatthecodec.home.presentation.model
+
+sealed interface ScreenMessage {
+    data object FileOpeningError : ScreenMessage
+    data object PermissionDeniedError : ScreenMessage
+    class ValueCopied(val value: String) : ScreenMessage
+}
