+++ app/src/main/java/com/javernaut/whatthecodec/home/presentation/model/VideoPage.kt
@@ -1,8 +1,8 @@
 package com.javernaut.whatthecodec.home.presentation.model
 
-import com.javernaut.whatthecodec.home.data.AudioStreamFeature
-import com.javernaut.whatthecodec.home.data.SubtitleStreamFeature
-import com.javernaut.whatthecodec.home.data.VideoStreamFeature
+import com.javernaut.whatthecodec.home.data.model.AudioStreamFeature
+import com.javernaut.whatthecodec.home.data.model.SubtitleStreamFeature
+import com.javernaut.whatthecodec.home.data.model.VideoStreamFeature
 import io.github.javernaut.mediafile.AudioStream
 import io.github.javernaut.mediafile.SubtitleStream
 import io.github.javernaut.mediafile.VideoStream
