+++ app/src/main/java/com/javernaut/whatthecodec/home/ui/audio/AudioPage.kt
@@ -6,7 +6,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import com.javernaut.whatthecodec.R
-import com.javernaut.whatthecodec.home.data.AudioStreamFeature
+import com.javernaut.whatthecodec.home.data.model.AudioStreamFeature
 import com.javernaut.whatthecodec.home.presentation.model.AudioPage
 import com.javernaut.whatthecodec.home.ui.stream.DisplayableStreamFeature
 import com.javernaut.whatthecodec.home.ui.stream.SimplePage
