+++ app/src/main/java/com/javernaut/whatthecodec/home/ui/screen/EmptyHomeScreen.kt
@@ -30,7 +30,7 @@ import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
 import com.javernaut.whatthecodec.R
 import com.javernaut.whatthecodec.compose.theme.WhatTheCodecTheme
-import com.javernaut.whatthecodec.home.presentation.ScreenMessage
+import com.javernaut.whatthecodec.home.presentation.model.ScreenMessage
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
 import kotlinx.coroutines.launch
